+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add TextEditor::AnyState const value.
 - Add TextEditor::add_key_binding() and remove_key_binding().
 - Fix for unicode fonts.
+- Pull FLTK menu fixes.
 - Update FLTK (removes experimental drivers).
 
 ## [1.2.28] - 2022-01-22